+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - `borax.calendars.festivals2`
   - `FestivalLibrary.iter_month_daytuples` 新增 `return_pos` 参数，可返回日期位置
 - `borax.htmls`
+  - `html_tag` 函数新增 width / height 参数
   - 修正 `html_tag` 函数style参数解析空值的BUG
 
 ## v3.5.4 (20220314)